requirements.txt
streamlit
librosa
praat-parselmouth
numpy
scipy
fpdf2
cryptography
PyPDF2
